git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/branches/wccp@1040383 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cluster/P_ClusterCache.h
@@ -581,7 +581,7 @@ struct ClusterVConnection: public ClusterVConnectionBase
   //  in the reply.
   virtual bool get_data(int id, void *data);    // backward compatibility
   virtual void get_http_info(CacheHTTPInfo **);
-  virtual int get_object_size();
+  virtual int64 get_object_size();
 
   // For VC(s) established via the HTTP version of OPEN_WRITE, additional
   //  data for the VC is passed in a second message.  This additional
@@ -942,7 +942,7 @@ ClusterFuncToQpri(int cluster_func)
  *									    *
  ****************************************************************************
  * Note: All of the following must be kept in sync with INKClusterRPCKey_t  *
- * 	 definition in ts/ts_private.h					    *
+ * 	 definition in ts/ts.h and ts/experimental.h			    *
  ****************************************************************************/
 
 /************************************************
